migrate all events to server.go file
